SimpleBig: Add typecasts to allow for int DATA_T

+++ examples/SimpleBig/kernels.cl
@@ -20,16 +20,16 @@ DATA_T complex(input_matrix_t *img) {
 
         for (int y = -NORTH + 1; y < SOUTH; y++) {
                 for (int x = -WEST + 1; x < EAST; x++) {
-                        DATA_T a = sin(getData(img, -1, 0));
-                        DATA_T b = native_sin(getData(img, 0, 1) * a);
+                        DATA_T a = sin((float)getData(img, -1, 0));
+                        DATA_T b = native_sin((float)getData(img, 0, 1) * a);
                         sum += getData(img, y, x) * a * (b / b);
                 }
         }
 
         DATA_T out = 0;
         for (int i = EAST; i >= -WEST; i--) {
-                sum *= cos((DATA_T)((int)getData(img, 0, i) % i) + sqrt(getData(img, 0, 0)));
-                out += sinpi(sum);
+                sum *= cos((DATA_T)((int)getData(img, 0, i) % i) + sqrt((float)getData(img, 0, 0)));
+                out += sinpi((float)sum);
                 out /= sum;
         }
